requirements.txt
tgcrypto
pyrogram
Pillow
plate
psycopg2-binary
pony
flask
apscheduler